fix: 인피니티 스크롤 문제 해결
