+++ src/main/java/com/example/portfolio/controller/ProjectController.java
@@ -25,9 +25,9 @@ import org.springframework.web.bind.annotation.RestController;
 import com.example.portfolio.dto.CategoryCreateDto;
 import com.example.portfolio.dto.CategoryDto;
 import com.example.portfolio.dto.CategoryUpdateDto;
-import com.example.portfolio.dto.PhotoListDto;
 import com.example.portfolio.dto.ProjectCreateDto;
 import com.example.portfolio.dto.ProjectDetailDto;
+import com.example.portfolio.dto.ProjectDetailPageDto;
 import com.example.portfolio.dto.ProjectListCustomDto;
 import com.example.portfolio.dto.ProjectListDto;
 import com.example.portfolio.dto.ProjectUpdateDto;
@@ -170,7 +170,7 @@ public class ProjectController {
 	}
 
 	@GetMapping("/photos/{id}")
-	public Slice<PhotoListDto> getPhotos(
+	public ProjectDetailPageDto getPhotos(
 			@PageableDefault( size = 10) Pageable pageable,
 			@PathVariable("id") Long projectId) {
 		return projectService.getPhotoList(pageable, projectId);
