+++ src/main/java/com/example/portfolio/dto/PhotoListDto.java
@@ -3,19 +3,8 @@ package com.example.portfolio.dto;
 public class PhotoListDto {
 	private Long id;
 	private String imageUrl; 
-	private String thumbnailUrl;
-	private String title;
 	
 	public PhotoListDto() {}
-
-	public PhotoListDto(Long id, String imageUrl, String thumbnailUrl, String title) {
-		super();
-		this.id = id;
-		this.imageUrl = imageUrl;
-		this.thumbnailUrl = thumbnailUrl;
-		this.title = title;
-	}
-
 
 	public PhotoListDto(Long id, String imageUrl) {
 		super();
@@ -38,22 +27,6 @@ public class PhotoListDto {
 
 	public void setImageUrl(String  imageUrl) {
 		this.imageUrl = imageUrl;
-	}
-
-	public String getThumbnailUrl() {
-		return thumbnailUrl;
-	}
-
-	public void setThumbnailUrl(String thumbnailUrl) {
-		this.thumbnailUrl = thumbnailUrl;
-	}
-
-	public String getTitle() {
-		return title;
-	}
-
-	public void setTitle(String title) {
-		this.title = title;
 	}
 	
 }
